+++ HolmSecurity/holm_security/asset_connector/device_assets.py
@@ -151,7 +151,7 @@ class HolmSecurityDeviceAssetConnector(AssetConnector):
             created_time=self._to_epoch(device.created),
             last_seen_time=self._to_epoch(device.last_sync),
             is_managed=True,
-            risk_score=device.risk_score if device.risk_score else None,
+            risk_score=device.risk_score,
             risk_level=risk_level,
             risk_level_id=risk_level_id,
         )
@@ -178,7 +178,7 @@ class HolmSecurityDeviceAssetConnector(AssetConnector):
             last_seen_time=self._to_epoch(asset.last_detected),
             is_managed=False,
             vendor_name=self.PRODUCT_NAME,
-            risk_score=asset.risk_score if asset.risk_score else None,
+            risk_score=asset.risk_score,
             risk_level=risk_level,
             risk_level_id=risk_level_id,
         )